==> 2. MFT processing.ipynb <==
import pandas as pd

## Importing MFT data

raw_mft_df = pd.read_excel("./data/MFT Returns by Zip 2019.xlsx", dtype={'Zip': object})

raw_mft_df.columns

#### Dropping columns with sensistive data

raw_mft_df.drop(axis='columns',
                labels=['#EITC returns', 'Average EITC Amount', 'Average Refund Amount', 'Average Age'],
                inplace=True)

#### Dropping unprocessable rows

print(raw_mft_df[pd.isnull(raw_mft_df['United Way'])].size); print(raw_mft_df.size)

Dropping any row that has a letter in the zip code or doesn't have an associated United Way. Because we will be filling unmatched zip codes with the closest zip code served by the same United Way, we can't use rows without a United Way. Luckily, the rows without a United Way appear to mostly be the rows with nonexistent or malformed zip codes.

This removes 3,619 rows from the data set, out of 125,545 total rows.

mft_df = raw_mft_df[~raw_mft_df.Zip.str.match("[a-zA-Z]") & ~pd.isnull(raw_mft_df['United Way'])].copy()

mft_df.to_csv("./data/mft_returns_2019.csv", index=False)

mft_df.sample(15)
