==> 3. Merging MFT to ACS relational file.ipynb <==
import pandas as pd

### Matching MFT to Census-known zip codes

In order to merge the MFT data (recorded at the zipcode level) with the IRS/ACS data (recorded at the county level), it is necessary to get convert from zip code to county as the summary level. To do this, we will use the Census' [zipcode to county relationship file](https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.2010.html).

mft_df = pd.read_csv("./data/mft_returns_2019.csv", dtype={'Zip': object})
census_rel_df = pd.read_csv("./data/zcta_county_rel_10.csv", dtype={'ZCTA5': 'object'})

Right outer merge will get all rows from the MFT dataset and any rows from the ACS dataset that match.

merged = pd.merge(census_rel_df, mft_df, how="outer", left_on="ZCTA5", right_on="Zip")

Sorting on zip because we'll be filling missing zip codes with their neighbors.

merged = merged.sort_values('Zip')

Forward filling missing zip codes. I.e. if we have two rows like

row#  | Zip     | United Way
:----|:--------|:----
1    | 48103   | United Way of Washtenaw County
2    | NA      | United Way of Washtenaw County

then the NA will become 48103. We're only filling within the same United Way, to make sure the nearest zip code isn't geographically super far away.

merged["filled_zip"] = merged.groupby('United Way') \
                             .apply(lambda group: group['ZCTA5'].fillna(method='ffill').fillna(method='bfill')) \
                             .reset_index(level=0, drop="True")

Creating a list of zip codes to use for spot checking.

sample_zips = merged['filled_zip'].dropna().sample(5)

# sorting by ZIP instead of Zip makes sure we have the ACS row first
# we don't need anything but the zip code, because we'll be joining this to the rel file again
acs_mft_df = merged.sort_values('ZCTA5') \
                    .groupby('filled_zip') \
                    .agg({'ZCTA5': 'first', 
                          'City': 'first',
                          'State': 'first',
                          'County': 'first',
                          'Org ID': 'first',
                          'United Way': 'first',
                          '#e-filed returns': 'sum'}) \
                     .reset_index(drop="True")

Spot-checking that the e-filed return numbers correctly aggregated.

merged[merged.filled_zip.isin(sample_zips)].groupby('filled_zip').agg({'#e-filed returns': 'sum'})

acs_mft_df[acs_mft_df.ZCTA5.isin(sample_zips)]

acs_mft_df.to_csv("./data/mft_county_rel.csv", index=False)
