==> 4. MFT at the county level.ipynb <==
# Aggregating MFT to the county level

In order to merge the MFT data (recorded at the zipcode level) with the IRS/ACS data (recorded at the county level), it is necessary to get convert from zip code to county as the summary level. To do this, we will use the Census' [zipcode to county relationship file](https://www.census.gov/geographies/reference-files/time-series/geo/relationship-files.2010.html).

import pandas as pd

### Merging relationship file and MFT data

rel_df = pd.read_csv("./data/zcta_county_rel_10.csv", dtype={'ZCTA5': object})
mft_df = pd.read_csv("./data/mft_county_rel.csv", dtype={'ZCTA5': object})

mft_county_rel = pd.merge(rel_df, mft_df, how='left')

Checking that we didn't leave anything behind, by looking for any rows where the zip code was in the MFT dataset but the United Way field is empty. There shouldn't be any rows where this is true.

mft_county_rel[mft_county_rel.ZCTA5.isin(mft_df.ZCTA5) & pd.isnull(mft_county_rel['United Way'])]

Everything looks good!

mft_county_rel.columns

### Aggregating to the county

Because zipcodes and counties are not co-terminous, some zipcodes are in multiple counties. The column `ZPOPPCT` in the zipcode to county relationship file denotes how much of the zipcode's population is in each county. For example, zip code `39573` stretches across 5 counties:

mft_county_rel[mft_county_rel.ZCTA5 == '39573'][['ZCTA5', 'City', 'State', 'County', 'STATE', 'COUNTY', 'ZPOPPCT']]

We want one county per zipcode, so we'll take only the row with the highest percentage of people. In the case of zipcode `38573`, that would be the last row (the one with a county FIPS code of `131`).

mft_county_rel = mft_county_rel.sort_values(['ZCTA5', 'COUNTY', 'ZPOPPCT']).drop_duplicates(['ZCTA5', 'COUNTY'], keep='last')

mft_county_rel.describe()

And, since some counties have lots of zipcodes, we need to get a total per county.

mft_by_county = mft_county_rel.groupby(['STATE', 'COUNTY']).agg(
    {
        '#e-filed returns': 'sum',
        'United Way': lambda gp: ', '.join(
            [x for x in set(gp.astype('str')) if x != 'nan']
        )
    }).reset_index()

### Reality Checks

Making sure all of the counties made it to `mft_county_rel` and that all e-filed returns made it to `mft_by_county

len(mft_county_rel.groupby(['STATE', 'COUNTY']).groups) == len(rel_df.groupby(['STATE', 'COUNTY']).groups)

mft_county_rel['#e-filed returns'].sum() == mft_by_county['#e-filed returns'].sum()

### Export

mft_by_county.to_csv("./data/mft_at_county_level.csv", index=False)
